🪒 Razor: Flatten Layer Lasagna and Remove Factory Factory
- Flattened deep, unnecessary nested directory hierarchies (`relvar-core/src/query`, `relvar-core/src/utils`, `relvar-core/src/storage_engine`) into top-level modules.
- Refactored `MockRelation` from a fluent builder "Factory Factory" to a direct generator function (`generate_mock_relation`).
- Ensured tests continue to pass and code conforms to KISS principle.
