+++ .jules/razor.md
@@ -21,3 +21,13 @@
 **Bloat:** Single-variant enum `RelationError` in `relvar-core/src/values/relation.rs` acting as a unit struct, with an unused variant `DuplicateTuple` and only one used variant `TypeMismatch`.
 **Cut:** Converted to unit struct `pub struct RelationError;` with `#[error("Tuple does not conform to relation type")]` directly. Removed the unused `DuplicateTuple` variant.
 **Saved:** Unnecessary enum matching, simplified error handling code significantly.
+
+## [Reduction]
+**Bloat:** Layer Lasagna (deep folder hierarchies with low file counts).
+**Cut:** Flattened `relvar-core/src/query/mod.rs` to `relvar-core/src/query.rs`, `relvar-core/src/utils/` to `relvar-core/src/utils.rs`, and `relvar-core/src/storage_engine/` to `relvar-core/src/storage_engine.rs`.
+**Saved:** Multiple unnecessary nested directories and improved codebase navigation speed.
+
+## [Reduction]
+**Bloat:** Factory Factory (Builders for objects taking only a few arguments).
+**Cut:** Converted `MockRelation` builder pattern in `relvar/src/experimental/mock.rs` to a simple standalone `generate_mock_relation` function.
+**Saved:** Unnecessary struct definition, builder chaining methods, and cognitive load for test creation.
