--- relvar-core/src/utils/mod.rs
@@ -1,7 +1,0 @@
-//! General utility functions and common structures.
-//!
-//! This module contains helper components that do not inherently belong to relational
-//! logic but are used across the codebase, such as memory bounds checking or recursion limits.
-pub(crate) mod recursion;
-
-pub use recursion::RecursionGuard;
